src/banco_de_dados/Conexao.java
package banco_de_dados;

import java.io.FileInputStream;
import java.sql.Connection;
import java.sql.DriverManager;
import java.util.Properties;

public class Conexao {
    
    public static Connection conectaMySql() {
        Connection conn = null;
        try {
        	Properties props = new Properties();
        	props.load(new FileInputStream("db.properties"));
            conn = DriverManager.getConnection(props.getProperty("url"),
            		    					   props);
            
        } catch (Exception e) {
            e.printStackTrace();
        }
        return conn;
    }

}
